+++ R/bigquery-execute.R
@@ -55,8 +55,10 @@ bqExecuteSql <- function(sql, ..., use.legacy.sql = bqUseLegacySql()) {
 
 
   if (!use.legacy.sql & length(params) > 0) {
-    cat("parameters applied to the template: \n")
-    message(jsonlite::toJSON(params, auto_unbox = TRUE, force = TRUE))
+    message(
+      "parameters applied to the template: \n",
+      jsonlite::toJSON(params, auto_unbox = TRUE, force = TRUE)
+    )
   } else {
     params <- NULL
   }
@@ -102,8 +104,10 @@ bqExecuteDml <- function(query, ...,
   )
 
   if (length(params) > 0) {
-    cat("parameters applied to the template: \n")
-    message(jsonlite::toJSON(params, auto_unbox = TRUE, force = TRUE))
+    message(
+      "parameters applied to the template: \n",
+      jsonlite::toJSON(params, auto_unbox = TRUE, force = TRUE)
+    )
   }
 
   ds <- bq_dataset(