+++ R/bigquery-table.R
@@ -37,7 +37,11 @@ bqDeleteTable <- function(table, dataset = bqDefaultDataset()) {
     dataset = dataset,
     table = table
   )
-  bq_table_delete(bt)
+  if (bq_table_exists(bt)) {
+    bq_table_delete(bt)
+  } else {
+    message("BigQuery table does not exist and will not be deleted: ", bt)
+  }
 }
 
 #' @rdname bqTable
@@ -152,4 +156,20 @@ bqCopyTable <- function(from, to, override = TRUE) {
   )
 
   return(bqTableExists(to))
+}
+
+
+#' Copies table through schema defintion
+#'
+#' @export
+#' @param from dataset object created by bigrquery::bq_table()
+#' @param to dataset object created by bigrquery::bq_table()
+bqCopyTableSchema <- function(from, to) {
+  meta <- bq_table_meta(from)
+  bq_table_create(
+    to,
+    fields = meta$schema$fields,
+    timePartitioning = meta$timePartitioning,
+    clustering = meta$clustering
+  )
 }